requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
drf-yasg
